+++ Akka.Persistence.Reminders/Reminder.Messages.cs
@@ -1,14 +1,13 @@
 ﻿#region copyright
 // -----------------------------------------------------------------------
 //  <copyright file="Reminder.Messages.cs" creator="Bartosz Sypytkowski">
-//      Copyright (C) 2017 Bartosz Sypytkowski <b.sypytkowski@gmail.com>
+//      Copyright (C) 2017-2020 Bartosz Sypytkowski <b.sypytkowski@gmail.com>
 //  </copyright>
 // -----------------------------------------------------------------------
 #endregion
 
 using System;
 using System.Collections.Immutable;
-using System.Text.RegularExpressions;
 using Akka.Actor;
 using Cronos;