+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+#### 0.3 March 09 2020
+* Updated Akka.Persistence and Akka.TestKit.Xunit to 1.4.1-rc1
+* Removed support for .NET 4.5 as no longer supported by Akka.Persistence
+* Bumped version to .NET Standard 2.0 to synchronized with Akka.Persistence
+* Bumped version of Tests to .NET Core 2.0 (minumum required by .NET Standard 2.0)
+* Replaced Akka.Configuration.ConfigurationFactory with Hocon.HoconConfigurationFactory
+
 #### 0.2 February 06 2019
 * Added support for cron expressions.